+++ src/server/src/projects/controllers/globalProjects.controller.ts
@@ -1,5 +1,15 @@
-// eslint-disable-next-line
-import { Controller, Get, ParseIntPipe, Query } from "@nestjs/common";
+// Not sure why, but eslint thinks these decorators are unused
+/* eslint-disable */
+import {
+  Controller,
+  Get,
+  ParseIntPipe,
+  Query,
+  ParseBoolPipe,
+  DefaultValuePipe
+} from "@nestjs/common";
+/* eslint-enable */
+
 import { Project } from "../entities/project.entity";
 import { ProjectsService } from "../services/projects.service";
 
@@ -12,8 +22,9 @@ export class GlobalProjectsController {
   @Get()
   async getAllGlobalProjects(
     @Query("page", ParseIntPipe) page = 1,
-    @Query("limit", ParseIntPipe) limit = 10
+    @Query("limit", ParseIntPipe) limit = 10,
+    @Query("completed", new DefaultValuePipe(false), ParseBoolPipe) completed = false
   ): Promise<Pagination<Project>> {
-    return this.service.findAllPublishedProjectsPaginated({ page, limit });
+    return this.service.findAllPublishedProjectsPaginated({ page, limit, completed });
   }
 }